Simplified functionality for os command utilities.
